extract uri constructor class into individual class
